Refactor HTTP middleware to separate package
