+++ server/middleware/middleware.go
@@ -1,4 +1,4 @@
-package gql
+package middleware
 
 import (
 	"context"
@@ -8,6 +8,7 @@ import (
 	"github.com/graphql-go/handler"
 	"github.com/jinzhu/gorm"
 	"github.com/maxbrain0/react-go-graphql/server/config"
+	"github.com/maxbrain0/react-go-graphql/server/logger"
 )
 
 type contextKey string
@@ -16,16 +17,18 @@ const contextKeyHeader = contextKey("header")
 const contextKeyDB = contextKey("db")
 const contextKeyAuth = contextKey("auth")
 
-// MiddlewareConfig holds references that will be accessed in middleware
-type MiddlewareConfig struct {
+var ctxLogger = logger.CtxLogger
+
+// Config holds references that will be accessed in middleware
+type Config struct {
 	GQLHandler *handler.Handler
 	DB         *gorm.DB
 	E          *casbin.Enforcer
-	AUTH       config.Auth
+	AUTH       *config.Auth
 }
 
 // HTTPMiddleware adds the request header to a graphql handler function
-func HTTPMiddleware(c MiddlewareConfig) http.Handler {
+func HTTPMiddleware(c Config) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		ctx := context.WithValue(r.Context(), contextKeyHeader, r.Header)
 		ctx = context.WithValue(ctx, contextKeyDB, c.DB)
@@ -35,32 +38,32 @@ func HTTPMiddleware(c MiddlewareConfig) http.Handler {
 }
 
 // GetHeader returns the header as a strgin
-func GetHeader(ctx context.Context) http.Header {
+func GetHeader(ctx context.Context) (http.Header, bool) {
 	header, ok := ctx.Value(contextKeyHeader).(http.Header)
 
 	if !ok {
 		ctxLogger.Warningln("Unable to get Header key in HTTPMiddleware")
 	}
 
-	return header
+	return header, ok
 }
 
 // GetDB retrieves gorm connection from context
-func GetDB(ctx context.Context) *gorm.DB {
+func GetDB(ctx context.Context) (*gorm.DB, bool) {
 	db, ok := ctx.Value(contextKeyDB).(*gorm.DB)
 
 	if !ok {
 		ctxLogger.Warningln("Unable to get DB key in HTTPMiddleware")
 	}
-	return db
+	return db, ok
 }
 
 // GetAuth loads a map with key strings to the oauth2 provider and values containing oauth2.config
-func GetAuth(ctx context.Context) config.Auth {
-	config, ok := ctx.Value(contextKeyAuth).(config.Auth)
+func GetAuth(ctx context.Context) (*config.Auth, bool) {
+	config, ok := ctx.Value(contextKeyAuth).(*config.Auth)
 
 	if !ok {
 		ctxLogger.Warningln("Unable to get Auth key in HTTPMiddleware")
 	}
-	return config
+	return config, ok
 }
